compute servrer: serve actual cocalc frontend app code.
- also found a cache header bug in hub express server (only impacts
  cocalc-docker so not a real issue).
