+++ src/compute/compute/lib/http-server.ts
@@ -7,8 +7,13 @@ import express from "express";
 import { createServer } from "http";
 import { getLogger } from "@cocalc/backend/logger";
 import type { Manager } from "./manager";
+import { path as STATIC_PATH } from "@cocalc/static";
+import { join } from "path";
+import { cacheShortTerm, cacheLongTerm } from "@cocalc/util/http-caching";
 
 const logger = getLogger("compute:http-server");
+
+const ENTRY_POINT = "app.html";
 
 export function initHttpServer({
   port = 5004,
@@ -26,8 +31,21 @@ export function initHttpServer({
 
   app.get("/", (_req, res) => {
     const files = manager.getOpenFiles();
-    res.send(`<h1>Compute Server</h1>  Open Files: ${files.join(", ")}`);
+    res.send(
+      `<h1>Compute Server</h1>  <a href="${join("/static", ENTRY_POINT)}">CoCalc App</a> <br/><br/> Open Files: ${files.join(", ")}`,
+    );
   });
+
+  app.use(
+    join("/static", ENTRY_POINT),
+    express.static(join(STATIC_PATH, ENTRY_POINT), {
+      setHeaders: cacheShortTerm,
+    }),
+  );
+  app.use(
+    "/static",
+    express.static(STATIC_PATH, { setHeaders: cacheLongTerm }),
+  );
 
   server.listen(port, host, () => {
     logger.info(`Server listening http://${host}:${port}`);
