+++ src/compute/compute/package.json
@@ -32,11 +32,13 @@
     "@cocalc/backend": "workspace:*",
     "@cocalc/compute": "link:",
     "@cocalc/jupyter": "workspace:*",
+    "@cocalc/static": "workspace:*",
     "@cocalc/sync": "workspace:*",
     "@cocalc/sync-client": "workspace:*",
     "@cocalc/sync-fs": "workspace:*",
     "@cocalc/terminal": "workspace:*",
     "@cocalc/util": "workspace:*",
+    "@types/ms": "^0.7.34",
     "@types/ws": "^8.5.9",
     "awaiting": "^3.0.0",
     "cookie": "^1.0.0",
