+++ src/compute/pnpm-workspace.yaml
@@ -9,3 +9,4 @@ packages:
   - "util"
   - "terminal"
   - "compute"
+  - "static"
