+++ src/packages/util/package.json
@@ -12,7 +12,8 @@
     "./sync/table": "./dist/sync/table/index.js",
     "./sync/editor/db": "./dist/sync/editor/db/index.js",
     "./licenses/purchase/*": "./dist/licenses/purchase/*.js",
-    "./redux/*": "./dist/redux/*.js"
+    "./redux/*": "./dist/redux/*.js",
+    "./http-caching": "./dist/http-caching.js"
   },
   "scripts": {
     "preinstall": "npx only-allow pnpm",
@@ -21,24 +22,15 @@
     "test": "pnpm exec jest",
     "prepublishOnly": "pnpm test"
   },
-  "files": [
-    "dist/**",
-    "bin/**",
-    "README.md",
-    "package.json"
-  ],
+  "files": ["dist/**", "bin/**", "README.md", "package.json"],
   "author": "SageMath, Inc.",
-  "keywords": [
-    "utilities",
-    "mathjax",
-    "markdown",
-    "cocalc"
-  ],
+  "keywords": ["utilities", "mathjax", "markdown", "cocalc"],
   "license": "SEE LICENSE.md",
   "dependencies": {
     "@ant-design/colors": "^6.0.0",
     "@cocalc/util": "workspace:*",
     "@types/debug": "^4.1.12",
+    "@types/ms": "^0.7.31",
     "async": "^1.5.2",
     "awaiting": "^3.0.0",
     "dayjs": "^1.11.11",
@@ -51,6 +43,7 @@
     "jsonic": "^1.0.1",
     "lodash": "^4.17.21",
     "lru-cache": "^7.18.3",
+    "ms": "2.1.2",
     "prop-types": "^15.7.2",
     "react-intl": "^6.7.0",
     "redux": "^4.2.1",
